Assignment 2/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-anthropic==0.1.3
python-dotenv==1.0.0
wikipedia==1.4.0
requests==2.31.0
google-search-results==2.4.2
beautifulsoup4==4.12.2
